fixed cases where indiv was already in correct side
if latency to enter correct side was zero, that means the individual was already in the correct side when the screens turned on. We cannot use that data, and it was messing up the mean latency to enter the correct side. I changed the zeros to NA's for both correct and incorrect latencies.